Move test_pipeline_is_not_stuck_with_components_with_only_defaults_as_first_components
